+++ src/apsg/helpers/__init__.py
@@ -1,12 +1,12 @@
 # -*- coding: utf-8 -*-
 
-from apsg.helpers._math import sind, cosd, tand, acosd, asind, atand, atan2d, sqrt2
-from apsg.helpers._helper import eformat
+from apsg.helpers._helper import eformat, is_jsonable
+from apsg.helpers._math import acosd, asind, atan2d, atand, cosd, sind, sqrt2, tand
 from apsg.helpers._notation import (
-    geo2vec_planar,
     geo2vec_linear,
-    vec2geo_planar,
+    geo2vec_planar,
     vec2geo_linear,
+    vec2geo_planar,
 )
 
 __all__ = (
@@ -21,6 +21,7 @@ __all__ = (
     "is_like_vec3",
     "is_like_matrix3",
     "eformat",
+    "is_jsonable",
     "geo2vec_planar",
     "geo2vec_linear",
     "vec2geo_planar",